[erp] Epson: поддержка копеек (refs #5263)

+++ erp-logics/src/main/java/lsfusion/erp/region/by/machinery/cashregister/fiscalepson/FiscalEpsonServiceInOutClientAction.java
@@ -26,7 +26,7 @@ public class FiscalEpsonServiceInOutClientAction implements ClientAction {
             FiscalEpson.init();
 
             FiscalEpson.openPort(comPort, baudRate);
-            FiscalEpson.inOut(sum.longValue());
+            FiscalEpson.inOut(sum.doubleValue());
             FiscalEpson.openDrawer();
 
         } catch (RuntimeException e) {
